Basic add lens activity added

+++ app/src/main/java/sfu/packages/cmpt276a2/Lens.java
@@ -27,7 +27,7 @@ public class Lens {
         return maximumAperture;
     }
 
-    public void setMaximumAperture(float maximumAperture) {
+    public void setMaximumAperture(double maximumAperture) {
         this.maximumAperture = maximumAperture;
     }
